requirements.txt
openai>=1.0.0
pinecone>=2.2.0
langchain>=0.0.300
langchain-community>=0.0.1
langchain-openai>=0.0.1
langchain-huggingface>=0.0.1
langchain-pinecone>=0.0.1
streamlit>=1.20.0
python-dotenv>=1.0.0
pytest>=7.0.0
tenacity>=8.0.0
numpy>=1.21.0
pandas>=1.5.0
typing-extensions>=4.0.0
instaloader>=4.9.6
sentence-transformers>=2.2.2
huggingface-hub>=0.16.0
torch>=2.0.0
transformers>=4.30.0

# New dependencies for enhanced features
PyPDF2>=3.0.0
python-docx>=0.8.11
requests>=2.28.0
beautifulsoup4>=4.11.0
plotly>=5.0.0
streamlit-option-menu>=0.3.0
streamlit-aggrid>=0.3.0
streamlit-ace>=0.1.1
langsmith>=0.1.0
